refactor: Move string method from AssertUtils to StringUtils
