+++ src/main/java/org/opentripplanner/transit/model/organization/Agency.java
@@ -1,7 +1,7 @@
 /* This file is based on code copied from project OneBusAway, see the LICENSE file for further information. */
 package org.opentripplanner.transit.model.organization;
 
-import static org.opentripplanner.util.lang.AssertUtils.assertHasValue;
+import static org.opentripplanner.util.lang.StringUtils.assertHasValue;
 
 import java.time.ZoneId;
 import java.util.Objects;
