--- src/main/java/org/opentripplanner/util/lang/AssertUtils.java
@@ -1,22 +1,0 @@
-package org.opentripplanner.util.lang;
-
-public class AssertUtils {
-
-  /** Private constructor to prevent creation of this utility class */
-  private AssertUtils() {}
-
-  /**
-   * Verify String value is NOT {@code null}, empty or only whitespace.
-   *
-   * @throws IllegalArgumentException if given value is {@code null}, empty or only whitespace.
-   */
-  public static String assertHasValue(String value) {
-    if (value == null || value.isBlank()) {
-      throw new IllegalArgumentException(
-        "Value can not be null, empty or just whitespace: " +
-        (value == null ? "null" : "'" + value + "'")
-      );
-    }
-    return value;
-  }
-}
