+++ src/main/java/org/opentripplanner/util/lang/StringUtils.java
@@ -11,4 +11,19 @@ public class StringUtils {
   public static boolean hasValue(String text) {
     return text != null && !text.isBlank();
   }
+
+  /**
+   * Verify String value is NOT {@code null}, empty or only whitespace.
+   *
+   * @throws IllegalArgumentException if given value is {@code null}, empty or only whitespace.
+   */
+  public static String assertHasValue(String value) {
+    if (value == null || value.isBlank()) {
+      throw new IllegalArgumentException(
+        "Value can not be null, empty or just whitespace: " +
+        (value == null ? "null" : "'" + value + "'")
+      );
+    }
+    return value;
+  }
 }
