+++ src/test/java/org/opentripplanner/util/lang/StringUtilsTest.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.util.lang;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
@@ -15,5 +16,18 @@ class StringUtilsTest {
     assertFalse(StringUtils.hasValue(""));
     assertFalse(StringUtils.hasValue(" "));
     assertFalse(StringUtils.hasValue("\n\t"));
+  }
+
+  @Test
+  public void assertValueExist() {
+    // Ok if any value
+    StringUtils.assertHasValue("a");
+
+    // Should fail for these values
+    var illegalValues = new String[] { null, "", " ", "\t", " \n\r\t\f" };
+
+    for (var it : illegalValues) {
+      assertThrows(IllegalArgumentException.class, () -> StringUtils.assertHasValue(it));
+    }
   }
 }
